Refactor the existing code
Change the layout just a little is fix a few small errors.
